Refactoring some functions for better reuse - bunch of functions to read data is moved to ner_utils
