+++ scripts/convert_data.py
@@ -12,99 +12,17 @@ import logging
 import os
 import glob
 import pathlib
-from collections import namedtuple
-import re
 from typing import Tuple
 
 from tqdm import tqdm
 from random import choices, shuffle
 from os.path import splitext
 
-BsfInfo = namedtuple('BsfInfo', 'id, tag, start_idx, end_idx, token')
+from ner_utils import convert_bsf
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.INFO)
 
-
-def format_token_as_beios(token: str, tag: str) -> list:
-    t_words = token.split()
-    res = []
-    if len(t_words) == 1:
-        res.append(token + ' S-' + tag)
-    else:
-        res.append(t_words[0] + ' B-' + tag)
-        for t_word in t_words[1: -1]:
-            res.append(t_word + ' I-' + tag)
-        res.append(t_words[-1] + ' E-' + tag)
-    return res
-
-
-def format_token_as_iob(token: str, tag: str) -> list:
-    t_words = token.split()
-    res = []
-    if len(t_words) == 1:
-        res.append(token + ' B-' + tag)
-    else:
-        res.append(t_words[0] + ' B-' + tag)
-        for t_word in t_words[1:]:
-            res.append(t_word + ' I-' + tag)
-    return res
-
-
-def convert_bsf(data: str, bsf_markup: str, converter: str = 'beios') -> str:
-    """
-    Convert data file with NER markup in Brat standoff Format to BEIOS or IOB format.
-
-    :param converter: iob or beios converter to use for document
-    :param data: tokenized data to be converted. Each token separated with a space
-    :param bsf_markup: Brat standoff Format markup
-    :return: data in BEIOS or IOB format https://en.wikipedia.org/wiki/Inside–outside–beginning_(tagging)
-    """
-
-    def join_simple_chunk(chunk: str) -> list:
-        if len(chunk.strip()) == 0:
-            return []
-        # keep the newlines, but discard the non-newline whitespace
-        tokens = re.split(r'(\n)|\s', chunk.strip())
-        # the re will return None for splits which were not caught in a group
-        tokens = [x for x in tokens if x is not None]
-        return [token + ' O' if len(token.strip()) > 0 else token for token in tokens]
-
-    converters = {'beios': format_token_as_beios, 'iob': format_token_as_iob}
-    res = []
-    markup = parse_bsf(bsf_markup)
-
-    prev_idx = 0
-    m_ln: BsfInfo
-    convert_f = converters[converter]
-    for m_ln in markup:
-        res += join_simple_chunk(data[prev_idx:m_ln.start_idx])
-
-        res.extend(convert_f(m_ln.token, m_ln.tag))
-        prev_idx = m_ln.end_idx
-
-    if prev_idx < len(data) - 1:
-        res += join_simple_chunk(data[prev_idx:])
-
-    return '\n'.join(res)
-
-
-def parse_bsf(bsf_data: str) -> list:
-    """
-    Convert multiline textual bsf representation to a list of named entities.
-
-    :param bsf_data: data in the format 'T9	PERS 778 783    токен'. Can be multiple lines.
-    :return: list of named tuples for each line of the data representing a single named entity token
-    """
-    if len(bsf_data.strip()) == 0:
-        return []
-
-    ln_ptrn = re.compile(r'(T\d+)\s(\w+)\s(\d+)\s(\d+)\s(.+?)(?=T\d+\s\w+\s\d+\s\d+|$)', flags=re.DOTALL)
-    result = []
-    for m in ln_ptrn.finditer(bsf_data.strip()):
-        bsf = BsfInfo(m.group(1), m.group(2), int(m.group(3)), int(m.group(4)), m.group(5).strip())
-        result.append(bsf)
-    return result
 
 
 def convert_bsf_in_folder(src_dir_path: pathlib.Path, dst_dir_path: pathlib.Path, converter: str = 'beios',
