+++ scripts/requirements.txt
@@ -1,3 +1,4 @@
 mitie
 tqdm
 sklearn
+requests
